Added a proximity check to service CreateTrail to prevent duplicates
I noticed while testing that the POST /trails endpoint would accept and create
multiple copies of the same trail by name. This is a consequence of me using
trail_id (uuid) as the key to my data store rather than trail name. I chose to
do this intentionally because there could be trails in different locations with
exactly the same name (as I discovered while looking for examples of actual trails
to use for testing). Rather than changing the key to the data store I instead chose
to implement a 25km radius check for any trails with exactly the same name before
creating a new trail. 25km is arbitrary and this adds some overhead to trail creation,
especially as the data store scales, but it seemed like a fair compromise for this
small project.