Added route for POST /login and protected other app endpoints with auth middeware
After this change all of the other endpoints besides /login and /health require a
the bearer roken generated by POST /login to be sent in with the header or they
will fail to authenticate.

The /login endpoint doesn't actually check the credentials that are sent in (yet...)
it just returns a token for any non-empty username/password.